Добавлен метод поиска по id денег. Изменён сервис для создания логов. Убрали выброс исключения, если сумма не найдена, её может и не быть при операции создания суммы.

+++ src/main/java/com/art/service/InvestorCashLogService.java
@@ -5,7 +5,6 @@ import com.art.model.InvestorsCash;
 import com.art.repository.InvestorCashLogRepository;
 import org.springframework.stereotype.Service;
 
-import javax.persistence.EntityNotFoundException;
 import java.util.List;
 
 /**
@@ -30,11 +29,7 @@ public class InvestorCashLogService {
      * @return найденная сумма
      */
     public InvestorCashLog findById(Long id) {
-        InvestorCashLog cashLog = investorCashLogRepository.findOne(id);
-        if (null == cashLog) {
-            throw new EntityNotFoundException("Не найдена сумма из лога операций с id = [" + id + "]");
-        }
-        return cashLog;
+        return investorCashLogRepository.findByCashId(id);
     }
 
     /**
@@ -61,7 +56,9 @@ public class InvestorCashLogService {
 
     public void delete(InvestorsCash cash) {
         InvestorCashLog cashLog = findById(cash.getId().longValue());
-        investorCashLogRepository.delete(cashLog);
+        if (null != cashLog) {
+            investorCashLogRepository.delete(cashLog);
+        }
     }
 
 }
